#4 datepicker로 diary의 date를 추가할수있게 수정, 다이어리 추가 할떄 데이터 추가 되도록 작업,  다이어리 리스트 디자인 변경, fontAwesome으로 기분 나타낼수있는 아이콘 보일수있게 작업

+++ client/src/Components/Diary/List.js
@@ -3,12 +3,26 @@ import React, { useState, useEffect, useLayoutEffect, useRef } from 'react';
 import { Link } from 'react-router-dom';
 import { useCookies } from 'react-cookie';
 import { connect } from 'react-redux';
-import moment from 'moment';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import {
+  faFaceAngry, // angry
+  faFaceGrin, // happy
+  faFaceSadTear, // sad
+  faFaceMeh, // soso
+  faFaceFrown, // 서운
+  faFaceGrinBeam, // very happy
+} from '@fortawesome/free-solid-svg-icons';
 
+import moment from 'moment';
 const List = ({ auth, common, handleUpdate, handleDelete }) => {
   const nowTime = moment('2022-11-04T12:45:21.461Z').format('YYYY-MM-DD');
-  console.log(nowTime);
 
+  const mood = [
+    { icon: faFaceAngry, color: 'red' },
+    { icon: faFaceFrown, color: 'red' },
+    { icon: faFaceMeh, color: 'orange' },
+    { icon: faFaceGrin, color: 'blue', icon: faFaceGrinBeam, color: 'blue' },
+  ];
   const [cookies, removeCookie] = useCookies(['x_auth']);
   const [isLoading, setIsLoading] = useState(false);
   const [lists, setLists] = useState([]);
@@ -24,9 +38,14 @@ const List = ({ auth, common, handleUpdate, handleDelete }) => {
   };
 
   const onHandleDeleteAll = () => {
+    if (!window.confirm('정말 삭제하시겠습니까? 복구 불가능 합니다.')) {
+      return;
+    }
+
     const config = { headers: { Authorization: cookies.x_auth } };
     axios.post(`${common.axiosUrl}/diary/deleteAll`, {}, config).then((res) => {
       if (res.data.status) {
+        setLists([]);
         alert('성공');
       } else {
         alert('권한 없음');
@@ -58,28 +77,20 @@ const List = ({ auth, common, handleUpdate, handleDelete }) => {
               {lists ? (
                 lists.map((item) => (
                   <div className='list-item' key={item._id}>
-                    <img src='/assets/image/bonobono_profile.jpg' alt='' />
                     <div className='info'>
-                      <Link to={`/detail/row/${item._id}`}>
-                        <p>
-                          <span>Title</span> {item.title}
-                        </p>
-                      </Link>
-                      <p className='regdate'>
-                        <span>RegDate</span>
-                        {moment(item.createdAt).format('YYYY-MM-DD HH:mm')}
-                      </p>
+                      <img src='/assets/image/bonobono_profile.jpg' alt='' />
+                      <div className='text'>
+                        <p className='regdate'>{moment(item.diaryDate).format('YYYY/MM/DD')}</p>
+                        <Link to={`/detail/row/${item._id}`}>
+                          <p>{item.title}</p>
+                        </Link>
+                      </div>
                     </div>
-
-                    {/* <div className='handleButton'>
-                <button onClick={() => onClickDelete(item.id)}>delete</button>
-                <button>
-                  <Link to={`/edit/${item._id}`}>update</Link>
-                </button>
-                <button>
-                  <Link to={`/detail/${item._id}`}>detail</Link>
-                </button>
-              </div> */}
+                    {/* 1 2 : red 3: orange 4 5: blue */}
+                    <FontAwesomeIcon
+                      icon={mood[parseInt(item.mood) - 1].icon}
+                      className={`list-svg ${mood[parseInt(item.mood) - 1].color}`}
+                    />
                   </div>
                 ))
               ) : (
